+++ keyboards/gggw/crosses/config.h
@@ -20,6 +20,7 @@
 #define SPI_MOSI_PIN GP3
 #define SPI_MISO_PIN GP0
 #define PMW33XX_CS_PIN GP6
+#define PMW33XX_LIFTOFF_DISTANCE 0x02
 
 #define MOUSE_EXTENDED_REPORT
 #define POINTING_DEVICE_DEBUG
