+++ keyboards/gggw/crosses/rules.mk
@@ -3,4 +3,5 @@ POINTING_DEVICE_DRIVER = pmw3360
 OLED_ENABLE = yes
 
 SRC += lib/achordion.c
+SRC += lib/crosses.c
 WEAR_LEVELING_DRIVER = rp2040_flash
